requirements-dev.txt
# Development dependencies (for linting, typing, and testing so Github Actions can run them)
mypy
pandas-stubs
types-requests
pytest
ruff
psycopg2-binary
types-psycopg2
